renamed CardsComponent in CardComponent
